Add $CLARA_HOME to the OpenCV default search paths
It's not recommended to install third party libraries into $CLARA_HOME,
but it may be simpler when quickly trying the demo services.

Add $CLARA_HOME to the list of install prefixes when OpenCV is searched,
to avoid a build error because OpenCV cannot be found, and the need of
setting the 'opencvDir' property for Gradle and 'OpenCV_ROOT' variable
for CMake, for something that may be common case.