actions: workaround for OpenCV Java library name
It seems that the OpenCV packages were renamed with a `d` suffix, due to
ABI breakage (see Changelog in [0]), and the shared libraries also have
been renamed with a `d` suffix.

But `org.opencv.core.Core.NATIVE_LIBRARY_NAME`, used by the Java
services to load the library, still returns the name without the added
suffix, so it cannot be found.

As a workaround, simply symlink `libopencv_java454d.so` to
`libopencv_java454.so` for now.

[0]: https://packages.ubuntu.com/source/jammy/opencv